Match and link lbsnap
