+++ src/MetroTRK/intrinsics.h
@@ -10,6 +10,7 @@ double __fabs(double);
 float __fabsf(float);
 double __frsqrte(double);
 int __rlwinm(int, int, int, int);
+int __rlwimi(int, int, int, int, int);
 
 void* __memcpy(void* dst, const void* src, unsigned long n);
